+++ src/fintrist_ds/__init__.py
@@ -4,3 +4,4 @@ Import all of the processing functions.
 from .engine import *
 from .catalog import CATALOG
 from .manage import *
+from .viz import *